+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Calendar.js - Change Log:
 
 ### Version 0.8.1:
+- Added public function "setEvents()", which adds an array of events and clears the original (fires onEventsSet, if available).
 - 
 
 ### Version 0.8.0: